fix(recurring): address Gemini review — currency mismatch, sort, dedup parseSubPayments
Fixes all 4 Gemini code review comments on PR #165:

1. HIGH: getPrimaryCurrencyAmount no longer falls back to a secondary-
   currency amount (would store e.g. USD value as COP). Returns
   existingAmount or 0 instead.
2. HIGH: Form amount field now uses primary-currency entry from
   sub_payments, not a cross-currency sum.
3. MEDIUM: upsertSubPayment sort now takes primaryCurrency param and
   sorts primary first, then alphabetical (matches migration ORDER BY).
4. MEDIUM: parseSubPayments extracted to shared utility at
   lib/utils/sub-payments.ts, replacing 3 inline copies.
